### 2019-07-18 #1580 @danielskatz
@fblavoie - please change manuscript.md to paper.md to whedon can find it
