### 2019-07-18 #1580 @danielskatz
@mgymrek - would you be willing to edit this submission?
