### 2019-07-19 a06bc59
issue resolved in show_model()
