### 2019-07-25 #1580 @kyleniemeyer
Hi @mgymrek, just another ping—we'd like to assign this submission to you.
