### 2019-07-25 495175d
Some fixes to paper.md
Fixed the affiliation block, and changed image commands to use local paths.
