+++ paper/paper.md
@@ -12,6 +12,7 @@ authors:
     affiliation: 1
 affiliations:
  - name: Universite de Sherbrooke, Faculty of Engineering, Department of Chemical & Biotechnological Engineering, 2500 boul. de l'Universite, Sherbrooke, QC, Canada, J1K 2R1
+   index: 1
 date: 15 July 2019
 bibliography: references.bib
 ---
@@ -20,11 +21,11 @@ bibliography: references.bib
 
 # Motivations
 
-Design of experiments (DOE) are very common, both in the research and the industrial fields \[@montgomery2017design]. Thanks to their intrinsic orthogonality \[@plackett1946design], DOEs make it possible to minimize the number of experiments to perform and lead to meaningful results using analyses of variance (ANOVA). 
+Design of experiments (DOE) are very common, both in the research and the industrial fields [@montgomery2017design]. Thanks to their intrinsic orthogonality [@plackett1946design], DOEs make it possible to minimize the number of experiments to perform and lead to meaningful results using analyses of variance (ANOVA). 
 
 Well-known commercial software allow users to make DOEs, to insert responses and to calculate ANOVA tables through simple user interfaces [@montgomery2017design]. These software have a lot of additional analysis options, but their main use remains the calculation of ANOVA tables. 
 
-Some packages written in open-source programming languages allow users to create DOEs, such as *pyDOE2* for Python [@baudin2015pydoe], or to calculate ANOVA, such as Python's *Statmodels* package [@perktold2017statsmodels]. However, no high-level packages allowing both the creation of DOEs and the calculation of ANOVA have been made available. Even commercial high-level programming interfaces, such as Matlab [@mathworks2005statistics;martinez2015computational], do not have this capability. In response to this lack, we have created a Python's package, named DOENOVA, whose primary objective is to ease the creation of DOEs followed by the calculation of ANOVA tables within the most widely adopted open-source and high-level programming interface, namely Python 3 [@oliphant2007python].
+Some packages written in open-source programming languages allow users to create DOEs, such as *pyDOE2* for Python [@baudin2015pydoe], or to calculate ANOVA, such as Python's *Statmodels* package [@perktold2017statsmodels]. However, no high-level packages allowing both the creation of DOEs and the calculation of ANOVA have been made available. Even commercial high-level programming interfaces, such as Matlab [@mathworks2005statistics; martinez2015computational], do not have this capability. In response to this lack, we have created a Python's package, named DOENOVA, whose primary objective is to ease the creation of DOEs followed by the calculation of ANOVA tables within the most widely adopted open-source and high-level programming interface, namely Python 3 [@oliphant2007python].
 
 
 
@@ -34,7 +35,7 @@ We developed a wrap-up to combine pyDOE2’s (version 1.2.0) capacity of creatin
 
 Baudin's pyDOE2 package [@baudin2015pydoe] allows users to create an experimental design according to a selected DOE type and a specific number of factors. This package was used in DOENOVA to create General full-factorial and Plackett-Burman factorial designs. For complete and fractional 2-level factorial designs, we have included our own codes into the DOENOVA package. We also added the capacity to add blocks and center points which is not possible with pyDOE2.
 
-Through its ordinary least squares (OLS) function, the Statsmodels' package, created by Perktold, Seabold and Taylor [@perktold2017statsmodels], allows to calculate N-way ANOVAs. From our experience, this package may be cumbersome when selecting specific factors or interactions of factors to include in the ANOVA as the nomenclature of the command lines was not intuitive. Moreover, Statsmodels cannot deal with blocks and center points.
[…]
