### 2019-07-25 #1580 @kyleniemeyer
Hi @fblavoie, please merge the PR I submitted that attempts to fix some things in your paper: https://github.com/fblavoie/doenova/pull/1
