### 2019-07-29 #1580 @labarba
👋  @fblavoie — I'm the Associate Editor in Chief this week, and I will be continuing to search for an editor to assign to your submission. But first, I need to ask you to add to the top of your paper:

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

(From our [Documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).)

As it is, I (a non-specialist) cannot quite parse what your software does, and that makes it hard for me to identify the right editor. Thanks!
