### 2019-07-30 c67b80e
Summary added in paper
