### 2019-07-30 154df3c
Update paper.md
