+++ paper/paper.md
@@ -50,7 +50,7 @@ The most important work in the development of the DOENOVA package was to automat
 
 In parallel to package development, we also developed a website ([doenova.tripli.ca](http://doenova.tripli.ca)) to provide package documentation. The website also contains an interactive designer in which the users can interactively modify the design parameters: DOE type, number of factors, number of replicates, number of blocks and number of center points. By changing a parameter, the webpage automatically calculates and displays the required total of runs and determines if an ANOVA can be calculated. A Python code, to copy in any Python console, is also generated according to the selected parameters. 
 
-![figs:layers](images/im0.png){:.image-caption}**Figure 1**. Web interactive designer generating Python codes. In the first step, the user can change DOE parameters. From any parameter change, the required number of runs is automatically updated and the possibility, or not, of calculating an ANOVA is shown. In the second step, the user can select which factors and interactions of factors to include in the model.
+![figs:layers](images/im0.png)**Figure 1**. Web interactive designer generating Python codes. In the first step, the user can change DOE parameters. From any parameter change, the required number of runs is automatically updated and the possibility, or not, of calculating an ANOVA is shown. In the second step, the user can select which factors and interactions of factors to include in the model.
 
 Once the design parameters have been selected and the experiments performed, the interactive webpage also proposes a form to select which factors, and their interactions, must be considered in the ANOVA model. Again, the Python code is automatically generated.
 
@@ -62,7 +62,7 @@ So far, DOENOVA can handle general full-factorial, Plackett-Burman and 2-level f
 
 By default, level values are coded in terms of -1, 1 and 0 (for center points) in 2-level and Plackett-Burman factorial designs. The user can replace these default values. For this, the interactive designer form page eases the task as it provides an easy-to-use interface to modify the default values from which the Python code is automatically generated.
 
-![](images/im1.png){:.image-caption}**Figure 2**. Python consoles and *xlsx* file. The user first copies and pastes the Python code generated by the interactive designer's first step. The user then exports the plan and fills out the result column in Libre Office Calc (or Microsoft Excel). The user then imports the filled out xlsx file and copies and pastes the Python code generated by the interactive designer's second step.
+![](images/im1.png)**Figure 2**. Python consoles and *xlsx* file. The user first copies and pastes the Python code generated by the interactive designer's first step. The user then exports the plan and fills out the result column in Libre Office Calc (or Microsoft Excel). The user then imports the filled out xlsx file and copies and pastes the Python code generated by the interactive designer's second step.
 
 
 
