### 2019-07-30 #1580 @fblavoie
Thanks for the follow-up.
I just added the summary to the paper.
If it can help to find an editor... an important part of the software features perform "basic under-grad" mathematics. People working in mathematics/statistics should be able to understand what the software does.
