### 2019-08-27 #1685 @whedon
[REVIEW]: DOENOVA: Simple DOE and ANOVA package for Python 3
**Submitting author:** @fblavoie (<a href="http://orcid.org/0000-0003-4574-1315">Francis B Lavoie</a>)
**Repository:** <a href="https://github.com/fblavoie/doenova" target ="_blank">https://github.com/fblavoie/doenova</a>
**Version:** 0.1.3
