+++ paper/references.bib
@@ -2,7 +2,8 @@
   title={Design and analysis of experiments},
   author={Montgomery, Douglas C},
   year={2017},
-  publisher={John wiley \& sons}
+  publisher={John wiley \& sons},
+  ISBN={1119386101}
 }
 
 @article{plackett1946design,
@@ -13,36 +14,39 @@
   number={4},
   pages={305--325},
   year={1946},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={https://doi.org/10.2307/2332195}
 }
 
 @misc{baudin2015pydoe,
   title={pyDOE},
   author={Baudin, M and others},
-  year={2015}
+  year={2015},
+  url={https://pythonhosted.org/pyDOE/}
 }
 
 @misc{perktold2017statsmodels,
-  title={Statsmodels: Statistics in python},
+  title={StatsModels: Statistics in Python},
   author={Perktold, Josef and Seabold, Skipper and Taylor, Jonathan and others},
   year={2017},
-  publisher={Oct}
+  url={https://www.statsmodels.org}
 }
 
-@book{mathworks2005statistics,
+@misc{mathworks2005statistics,
   title={Statistics Toolbox: For Use with Matlab: User's Guide, Version 5},
   author={MathWorks, Inc},
   year={2005},
-  publisher={MathWorks}
+  publisher={MathWorks},
+  url={http://cda.psych.uiuc.edu/matlab_pdf/stats.pdf}
 }
 
 @book{martinez2015computational,
   title={Computational statistics handbook with MATLAB},
   author={Martinez, Wendy L and Martinez, Angel R},
   year={2015},
-  publisher={Chapman and Hall/CRC}
+  publisher={Chapman and Hall/CRC},
+  ISBN={9781466592735}
 }
-
 
 @article{oliphant2007python,
   title={Python for scientific computing},
@@ -52,5 +56,6 @@
   number={3},
   pages={10--20},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={https://doi.org/10.1109/MCSE.2007.58}
 }
