### 2019-09-29 fblavoie/doenova#2 @klmedeiros
JOSS DOENOVA review missing issues
hi! I'm kaelen and i'm completing the JOSS review of DOENOVA. I think, based on https://github.com/openjournals/joss-reviews/issues/1685, that you're missing a few things in the documentation, namely:

- automated tests
- community guidelines

I also couldn't find these on the accompanying website (http://doenova.tripli.ca/index.html, which i love, by the way!!)

i'm especially sure i'm missing the tests somewhere, but i've scoured the repo and am coming up empty.
