### 2019-09-29 #1685 @klmedeiros
i'm close to done on this, but i have a few issues blocking acceptance, posted here: https://github.com/fblavoie/doenova/issues/2

also, is it an issue that some of what is listed in the review lives on the accompanying website, not in the repo?
