### 2019-10-01 e5fefb6
Contributing, Tests, Website added; Correction in __init__
