+++ tests/test_anova3.py
@@ -1,0 +1,24 @@
+from doenova import anovan
+
+
+X = [[20, -1, 0],
+     [20, -1, 5],
+     [40, -1, 0],
+     [40, -1, 5],
+     [60, -1, 0],
+     [60, -1, 5],
+     [20,  1, 0],
+     [20,  1, 5],
+     [40,  1, 0],
+     [40,  1, 5],
+     [60,  1, 0],
+     [60,  1, 5]]
+
+y = [5.6, 3.4, 7.6, 2.3, 6.4, 7.1,
+     3.4, 1.2, 1.5, 9.8, 5.7, 3.4]
+
+sel = [[1, 0, 0], # A
+       [0, 1, 1], # BC
+       [1, 1, 1]] # ABC
+ 
+anova_comps = anovan(y,X,sel)
