### 2019-10-01 #1685 @fblavoie
@klmedeiros ,
Contributing instructions and test scripts were indeed missing. They were added. I have also uploaded the website on the github repo.
