### 2019-10-17 #1685 @usethedata
@fblavoie -- in your references.bib file, you need just the doi, not the url prefix.  For example, with the plackett1946design reference, you have `doi={https://doi.org/10.2307/2332195}`  This needs to be `doi={10.2307/2332195}`.  Can you fix this and the oliphant2007python reference?
