### 2019-11-02 #1685 @danielskatz
👋 @usethedata - how is this going?
