### 2019-11-03 #1685 @usethedata
@fblavoie I gave you incomplete/inaccurate information on fixing the DOI in the bibliography.  I've submitted a pull request to your repo that should fix the issue.  I also suggest closing the issue you have open from klmedeiros review. Once those are done, then I believe the paper is ready for me to request acceptance.

My apologies for the delays on this end.  I've had some family needs that have taken a lot of time.
