### 2019-11-24 #1685 @usethedata
@fblavoie checking in -- did you have a chance to look at the the pull request I submitted on your repo to correct the bibliography?  Thanks.
