### 2020-02-01 #1685 @arfon
@fblavoie - please merge this pull request that makes a number of small changes to your paper: https://github.com/fblavoie/doenova/pull/4

When you have done this could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
