### 2020-02-20 #1685 @arfon
It's now more than four months since [we last heard](https://github.com/openjournals/joss-reviews/issues/1685#issuecomment-537184269) from the author (@fblavoie). If there's no response by March 1st (which will be 5 months since last contact) then I will close this issue and withdraw the paper on their behalf, assuming that they have lost interest in pursuing this publication with JOSS.
