### 2020-03-01 #1685 @arfon
@klmedeiros, @usethedata - thanks again for your reviews here. Unfortunately as the author has been absent now for 5 months I'm think the only reasonable conclusion I can make is that the author is no-longer interested in pursuing this paper in JOSS. As such, I'm going to go ahead and withdraw the submission on their behalf.
